Moved task components in one folder
