+++ src/app/app.module.ts
@@ -8,11 +8,11 @@ import { AppComponent } from './app.component';
 import { AuthorizationInterceptor } from './interceptors/authorization.interceptor';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { AnonymousComponent } from './components/anonymous/anonymous.component';
-import { NewTaskComponent } from './components/new-task/new-task.component';
+import { NewTaskComponent } from './components/task/new-task/new-task.component';
 import { FormsModule } from '@angular/forms';
 import { NotInPastDirective } from './validators/not-in-past.directive';
-import { TaskDetailsComponent } from './components/task-details/task-details.component';
-import { EditTaskComponent } from './components/edit-task/edit-task.component';
+import { TaskDetailsComponent } from './components/task/task-details/task-details.component';
+import { EditTaskComponent } from './components/task/edit-task/edit-task.component';
 
 @NgModule({
   declarations: [AppComponent, DashboardComponent, AnonymousComponent, NewTaskComponent, NotInPastDirective, TaskDetailsComponent, EditTaskComponent],
