+++ src/app/components/task/edit-task/edit-task.component.scss
@@ -1,4 +1,4 @@
-@import "../../../colours.scss";
+@import "src/colours.scss";
 
 .panel {
   border-radius: 5px;
