+++ src/app/components/task/new-task/new-task.component.scss
@@ -1,4 +1,4 @@
-@import "../../../colours.scss";
+@import "src/colours.scss";
 
 .panel {
   border-radius: 5px;
